Updates + tests
