+++ api/barriers/serializers/__init__.py
@@ -1,5 +1,4 @@
 from .barriers import BarrierDetailSerializer, BarrierListSerializer  # noqa
-from .csv import BarrierCsvExportSerializer  # noqa
 from .data_workspace import DataWorkspaceSerializer  # noqa
 from .progress_updates import ProgressUpdateSerializer  # noqa
 from .public_barriers import PublicBarrierSerializer, PublishedVersionSerializer  # noqa
